organize things into files
the code runs, the animation is produced, seems like agents move too fast.
